Renamed CandidateInterface::CoursesRecommender to CandidateCoursesRecommender
