+++ app/components/support_interface/application_choice_component.rb
@@ -159,7 +159,7 @@ module SupportInterface
       return unless application_choice.submitted?
 
       key = 'Recommended courses'
-      recommended_courses_url = CandidateInterface::CoursesRecommender.recommended_courses_url(candidate: application_choice.candidate, locatable: application_choice.provider).presence
+      recommended_courses_url = CandidateCoursesRecommender.recommended_courses_url(candidate: application_choice.candidate, locatable: application_choice.provider).presence
       value = govuk_link_to(recommended_courses_url, class: 'govuk-link govuk-link--no-visited-state', target: '_blank', rel: 'noopener') do
         "Recommended courses (based on #{application_choice.provider.name} location)"
       end
