+++ app/views/support_interface/application_forms/show.html.erb
@@ -5,13 +5,13 @@
 <%= render SupportInterface::ApplicationSummaryComponent.new(application_form: @application_form) %>
 
 <div class="govuk-button-group">
-  <% if (recommended_courses_url = CandidateInterface::CoursesRecommender.recommended_courses_url(candidate: @application_form.candidate, locatable: @application_form).presence) %>
+  <% if (recommended_courses_url = CandidateCoursesRecommender.recommended_courses_url(candidate: @application_form.candidate, locatable: @application_form).presence) %>
     <div>
       <%= govuk_link_to 'Recommended Courses (based on Candidate location)', recommended_courses_url, class: 'govuk-link govuk-link--no-visited-state', target: '_blank', rel: 'noopener' %>
     </div>
   <% end %>
 
-  <% if (recommended_courses_url = CandidateInterface::CoursesRecommender.recommended_courses_url(candidate: @application_form.candidate).presence) %>
+  <% if (recommended_courses_url = CandidateCoursesRecommender.recommended_courses_url(candidate: @application_form.candidate).presence) %>
     <div>
       <%= govuk_link_to 'Recommended Courses (without a location)', recommended_courses_url, class: 'govuk-link govuk-link--no-visited-state', target: '_blank', rel: 'noopener' %>
     </div>
